// src/types.d.ts
declare module "@/types" {
    interface Tab {
        id: string;
        url: string;
        title: string;
    }

    // This allows TypeScript to pick up the magic constants that's auto-generated by Forge's Vite
    // plugin that tells the Electron app where to look for the Vite-bundled app code (depending on
    // whether you're running in development or production).
    const MAIN_WINDOW_VITE_DEV_SERVER_URL: string;
    const MAIN_WINDOW_VITE_NAME: string;

    // Preload types
    interface ThemeModeContext {
        toggle: () => Promise<boolean>;
        dark: () => Promise<void>;
        light: () => Promise<void>;
        system: () => Promise<boolean>;
        current: () => Promise<"dark" | "light" | "system">;
    }
    interface ElectronWindow {
        minimize: () => Promise<void>;
        maximize: () => Promise<void>;
        close: () => Promise<void>;
    }

    interface Window {
        themeMode: ThemeModeContext;
        electronWindow: ElectronWindow;
    }
}